audio: fixed dtv audio pts check issue [1/1]
PD#SWPL-146315

Problem:
Enable AD and subtitle,play the BBC ONE
service on BBC MUX UH.ts long time,
A/V out of sync Frequency:1/1pcs/48H

Solution:
need use uint64_t to calc instead of
unsigned int

Verify:
sc2

+++ input/include/device_patch.h
@@ -158,7 +158,7 @@ struct aml_audio_patch
     unsigned char dtv_NchOriginal;
     unsigned char dtv_lfepresent;
     unsigned int dtv_first_apts;
-    unsigned int dtv_pcm_wrote;
+    uint64_t  dtv_pcm_wrote;
     unsigned int dtv_pcm_readed;
     unsigned int dtv_decoder_ready;
     unsigned int input_thread_created;
